compute FactSet timestamp/s from filenames (#125)
depends on #99 

This PR computes the FactSet timestamp/s from the filenames of all input
FactSet data files, and adds that info to the manifest. Note that it
would be possible to use FactSet data files from different timestamps...
in which case all unique timestamps would be added to the manifest as an
array. In the future, e.g. in #118, we could consider giving an error or
warning if all of these timestamps do not match.