config: update 2023Q4 scenario preparation config path
The initial scenario preparation configuration bump that was committed in #236 and #237 pointed to a scenario preparation output that was missing crucial data. 

The data was missing due to a bug that was introduced in:
https://github.com/RMI-PACTA/pacta.scenario.data.preparation/pull/61

and later solved by:
https://github.com/RMI-PACTA/pacta.scenario.data.preparation/pull/63

The new scenario preparation outputs, in the directory: 
`2023Q4_20240701_T105302Z`

which can be found here: https://pactadatadev.file.core.windows.net/workflow-scenario-preparation-outputs/2023Q4_20240701_T105302Z

contain all the expected output datasets.